requirements.txt
deprecation==2.1.0
packaging==25.0
PyYAML==6.0.3
splunk-sdk==2.1.1
